+++ CHANGELOG.md
@@ -142,3 +142,5 @@ significant modifications will be credited to OpenTelemetry Authors.
 ([#583](https://github.com/open-telemetry/opentelemetry-demo/pull/583))
 * Change ZipCode data type from int to string
 ([#587](https://github.com/open-telemetry/opentelemetry-demo/pull/587))
+* Pass product's `categories` as an input for the Ad service
+([#600](https://github.com/open-telemetry/opentelemetry-demo/pull/600))